+++ API/ManagementAPI/ManagementAPI.IntegrationTests.Specflow/Common/TokenType.cs
@@ -1,0 +1,8 @@
+﻿namespace ManagementAPI.IntegrationTests.Specflow.Common
+{
+    public enum TokenType
+    {
+        Client,
+        Password
+    }
+}
