+++ API/ManagementAPI/ManagementAPI.IntegrationTests.Specflow/ManagementAPI.IntegrationTests.Specflow.csproj
@@ -138,6 +138,8 @@
       <DependentUpon>GetClubConfiguration.feature</DependentUpon>
     </Compile>
     <Compile Include="Common\Setup.cs" />
+    <Compile Include="Common\TokenResponse.cs" />
+    <Compile Include="Common\TokenType.cs" />
     <Compile Include="Player\PlayerSteps.cs" />
     <Compile Include="Player\RegisterPlayer.feature.cs">
       <AutoGen>True</AutoGen>
