+++ API/ManagementAPI/ManagementAPI.Service/Common/CustomClaims.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace ManagementAPI.Service.Common
+{
+    public class CustomClaims
+    {
+        /// <summary>
+        /// The club identifier
+        /// </summary>
+        public const String ClubId = "ClubId";
+
+        /// <summary>
+        /// The player identifier
+        /// </summary>
+        public const String PlayerId = "PlayerId";
+    }
+}
