+++ API/ManagementAPI/ManagementAPI.Service/Common/RoleNames.cs
@@ -1,0 +1,22 @@
+﻿using System;
+
+namespace ManagementAPI.Service.Common
+{
+    public class RoleNames
+    {
+        /// <summary>
+        /// The club administrator
+        /// </summary>
+        public const String ClubAdministrator = "CLUB ADMINISTRATOR";
+
+        /// <summary>
+        /// The match secretary
+        /// </summary>
+        public const String MatchSecretary = "MATCH SECRETARY";
+
+        /// <summary>
+        /// The player
+        /// </summary>
+        public const String Player = "PLAYER";
+    }
+}
